* Added nested tween * Added possibility to build with browserify
