+++ .gitignore
@@ -1,3 +1,4 @@
+node_modules
 build/tween.min.js
 npm-debug.log
 test.html
